Added interfaces
