+++ app/src/main/java/edu/washington/cs/ubicomplab/rdt_reader/fragments/SettingDialogFragment.java
@@ -23,6 +23,7 @@ import android.widget.RadioGroup;
 import android.widget.SeekBar;
 
 import edu.washington.cs.ubicomplab.rdt_reader.R;
+import edu.washington.cs.ubicomplab.rdt_reader.interfaces.SettingDialogListener;
 import edu.washington.cs.ubicomplab.rdt_reader.utils.Constants;
 
 public class SettingDialogFragment extends DialogFragment implements RadioGroup.OnCheckedChangeListener {
@@ -47,10 +48,6 @@ public class SettingDialogFragment extends DialogFragment implements RadioGroup.
         } else  if (i == R.id.bmButton) {
             Constants.LANGUAGE = "bm";
         }
-    }
-
-    public interface SettingDialogListener {
-        void onClickPositiveButton();
     }
 
     @Override
