+++ app/src/main/java/edu/washington/cs/ubicomplab/rdt_reader/interfaces/SettingDialogListener.java
@@ -1,0 +1,5 @@
+package edu.washington.cs.ubicomplab.rdt_reader.interfaces;
+
+public interface SettingDialogListener {
+    void onClickPositiveButton();
+}
